# ferrox-cp/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ferrox Control Plane API",
    "description": "Control plane for the Ferrox LLM gateway: API client administration, usage and budgets, signing-key rotation and the audit log (admin key required), plus the public token exchange, JWKS and health endpoints.",
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/audit": {
      "get": {
        "tags": [
          "Audit"
        ],
        "summary": "`GET /api/audit`",
        "description": "Lists audit log entries with optional filters.\nQuery parameters: `client_id`, `event`, `since`, `limit`, `offset`.",
        "operationId": "list_audit",
        "parameters": [
          {
            "name": "client_id",
            "in": "query",
            "description": "Only entries for this client.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid",
              "nullable": true
            }
          },
          {
            "name": "event",
            "in": "query",
            "description": "Only entries of this event type (e.g. `token_issued`).",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "since",
            "in": "query",
            "description": "Only entries created at or after this time.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size (capped at 1000).",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 100,
              "maximum": 1000
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of entries to skip.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Audit log entries, newest first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditEntry"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid admin key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Database failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "description": "The control plane's error envelope: `{\"error\": ..., \"message\": ...}`.\n\nSchema-only: admin handlers and the admin-key middleware build it as a JSON\nvalue, and `POST /token` uses `crate::handlers::token::ErrorResponse`; all of\nthem serialize to this shape.",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Short error code: an HTTP status code (e.g. `\"404\"`, most admin handler\nerrors) or a category (e.g. `unauthorized`, `internal_error`).",
            "example": "404"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message.",
            "example": "client not found"
          }
        }
      },
      "AuditEntry": {
        "type": "object",
        "description": "A single entry in the immutable audit log.",
        "required": [
          "id",
          "event",
          "created_at"
        ],
        "properties": {
          "client_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "event": {
            "type": "string",
            "description": "Event type: `token_issued`, `client_created`, `client_revoked`,\n`key_rotated`, `budget_exceeded`, or any other recorded string.",
            "example": "token_issued"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "metadata": {
            "type": "object",
            "description": "Free-form, event-specific JSON details.",
            "nullable": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "admin_auth": {
        "type": "http",
        "scheme": "bearer",
        "description": "The static admin key configured as `CP_ADMIN_KEY`."
      },
      "client_key_auth": {
        "type": "http",
        "scheme": "bearer",
        "description": "A client API key (`sk-cp-...`) as returned by `POST /api/clients`."
      }
    }
  },
...
}